refactor(lyrics): extract utils, hooks, components, config from LyricsContent
- utils/findCurrentIndex.ts: binary search for current lyric line
- components/KaraokeSyllableLine.tsx: karaoke syllable rendering
- hooks/useLyricsSettings.ts: karaoke, clock, glow settings
- hooks/useBeijingClock.ts: Beijing time clock (minute-boundary updates)
- hooks/useAutoIdle.ts: auto idle on music stop or empty lyrics
- hooks/useCurrentLyric.ts: current lyric index and line data
- config/lyricsConstants.ts: store key constant
- LyricsContent.tsx: simplified from 262 to 126 lines
