+++ src/renderer/components/states/lyrics/utils/findCurrentIndex.ts
@@ -1,0 +1,45 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file findCurrentIndex.ts
+ * @description 二分查找当前歌词行索引
+ * @author 鸡哥
+ */
+
+import type { SyncedLyricLine } from '../../../../store/types';
+
+/**
+ * 二分查找当前播放位置对应的歌词行索引。
+ * @param lyrics - 已同步的歌词行数组。
+ * @param posMs - 当前播放位置（毫秒）。
+ * @returns 当前行索引，未命中返回 -1。
+ */
+export function findCurrentIndex(lyrics: SyncedLyricLine[], posMs: number): number {
+  if (lyrics.length === 0 || posMs < lyrics[0].time_ms) return -1;
+  let lo = 0;
+  let hi = lyrics.length - 1;
+  while (lo < hi) {
+    const mid = (lo + hi + 1) >> 1;
+    if (lyrics[mid].time_ms <= posMs) lo = mid;
+    else hi = mid - 1;
+  }
+  return lo;
+}
